Enable game detail navigation from Player Absences report
Changes:
- Extracted GameDetailModal into separate reusable component
- Updated PlayerAbsencesReport to open game detail modal when clicking absence dates
- Updated GameEntry to import GameDetailModal instead of inline definition
- Removed duplicate GameDetailModal code from GameEntry.jsx

User flow:
1. View Player Absences report
2. Click on any absence date badge
3. Game detail modal opens showing full game information
4. View all player data, pitch counts, positions, and violations for that game

The modal is now shared between GameEntry and PlayerAbsencesReport,
eliminating code duplication and ensuring consistent game detail display.
